+++ CHANGES.md
@@ -4,6 +4,8 @@
 Changes in 7.2
 --------------
 - Add Grade Level breakdown in StudentFieldBreakdown.php
+- Add link to Student Info in AddDrop.php
+- Limit students to User schools in AddDrop.php
 
 Changes in 7.1.2
 ----------------